Working on converting the Javascript to a JQuery UI widget.
I don't think the code is that nice, but so far, it seems to be working.
The advantage is that all the necessary table is generated using JQuery
rather than PHP.

Working on a function to set selected times by default when constructing the
widget and on a getTime function that will return a Json string of all the
times that the user selected.