MAYOR BUG FIX: all participating children visible in /selfservice/welcome
A mayor bug resulted in showing a list of all children in the participants table
when the page selfservice/welcome was called without passing through authentication.
This is because the function current_email() then returns false (no current e-mail could
be found). This false is then passed to an SQL query which somehow selects EVERY row from the
table when "...where `email` = FALSE" is used.

The issue was solved by checking if current_email() returns false and, if so, redirecting to the
selfservice page, without loading any data from the database.

Other updates include:
- Selfservice is completely ready for multilang. Some changes are made in the language files as well, but those await approval
- Email language also depends on the language that the user has selected in their settings
- The "if this url does not work" link now also includes the base url, because otherwise it's a bit confusing for the user
- Added a link that refers to the participation registration window
- Now checks if the used email address actually exists in the database