Custom Babylab mail library added (location: application/libraries/Mail.php).
Each of the e-mails that is sent is now being sent through this babylab mail library.
The library works roughly the same as the CI email libary. It builds the mail and when
it is finished, you can send it with $this->mail->send().

However, the library does more.

1) It formats all emails in the default template. The mail->message(arg) arg only takes
raw html, without <html> or <body> tags and without a greeting, ending or disclaimer
2) By default the custom greeting "Beste" or "Dear" (dependent on the language) is used.
How the recipient is adressed can be set with mail->to_name(name). This will appear, together
with the greeting, in bold on the top of the mail. The custom greeting can be overridden, but it
makes it less work to format a decent email
3) The disclaimer is automatically added in the footer. This can be changed in the language files
4) By default, the email is sent from the babylab email and name. This cannot be overridden (as yet)
5) By default, the email is ended with the default "Hartelijke groet,
Het Babylab team van de Universiteit Utrecht". This can be overridden or deleted
6) The 'to' and 'bcc' fields automatically send to the override address if that should happen. No need
to set this every time an email is sent!
7) The subject line now only requires the actual subject. This is automatically appended to the "Babylab Utrecht: " tag

In this update, a few bugs were fixed as well:
- In users.php:195 a call was made to a view that required the 'referrer' parameter, which was not given to it. This is fixed
- call.php:3xx: When an email with an attachment was sent, the result was an empty email on my machine. It turns out this was because
the file could not be found. Whenever this happens from the call.php file now, the attachment is not added and a new comment is created
in the database about which file is missing from the server and who missed out on receiving that file as an attachment

Encountered problems that were not or just sort-of fixed:
- Call.php:367 	A call was done to lang('testinvite_added') but that key did not exist in any of the language files. I added the language
string for that key. Then it still gave the error "too few arguments" which didn't make sense, but I noticed the error went away as soon as
I printed the lang('testinvite_added') before using it. My solution was making a new variable that contains the language string and then use
the variable in the sprintf function. Why this should work, I have no iea
- call.php:144  testinvites[0] results in an error (no testinvites[0] exists) if you make an appointment with someone, wait for the confirmation
message, then click the back arrow in the browser, then don't change anything in the form and click "confirm" again. This is because no new test-invites
are made then, meaning the for-loop in create_test_invites doesn't run and no information is returned.
- call.php:~140  The flashdata created in the forloop in the function that is called somewhere over there doesn't turn up. No idea why.
- The datehelper class uses server language for the time strings. This looks a bit ugly in the emails